# service/src/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Mage API
  description: Mage API
  contact:
    name: Mage Support
    email: mage@nga.mil
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: 6.0.1
servers: []
security:
  - mageToken: []
...
paths:
...
  /auth/{authenticationStrategy}/signin:
    post:
      tags: [ Authentication & Authorization ]
      summary: Authentication
      description: Authenticate a user using a given strategy.
      operationId: authenticate
      security:
        []
      parameters:
        - name: authenticationStrategy
          in: path
          description: Authentication strategy.
          required: true
          schema:
            type: string
            enum: [ local, ldap, saml, oauth, openidconnect ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserIn' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/UserIn' }
      responses:
        200:
          description: Successfully signed in.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Authentication' }
  /auth/token:
    post:
      tags: [ Authentication & Authorization  ]
      summary: Device Authorization
      description: Authorize a device.  If device does not exist, it may also be created via this method.
      operationId: authorize
      security:
        - authenticationToken: []
      parameters:
        - name: uid
          in: query
          description: Id of the device that requires provisioning
          required: true
          schema: { $ref: '#/components/schemas/Device/properties/id' }
        - name: name
          in: query
          description: Name of the device
          schema: { $ref: '#/components/schemas/Device/properties/name' }
        - name: description
          in: query
          description: Description of the device
          schema: { $ref: '#/components/schemas/Device/properties/description' }
        - name: appVersion
          in: query
          description: Application version of the device (required if creating a new device)
          schema: { $ref: '#/components/schemas/Device/properties/appVersion' }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceIn' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/DeviceIn' }
      responses:
        200:
          description: Successful device authorization, Mage api token is returned.
        401:
          description:  User is not authorized.  Typically this means you need to authenticate `/auth/local/signin`.
        403:
          description: Failure to provision the device.  This typically means the device needs to be registered.
...
  /api/next-users/search:
    get:
      tags: [ User ]
      description: Search for users. The client must have `READ_USER` permission.
      summary: User search.
      operationId: userSearch
      parameters:
        - in: query
          name: term
          schema:
            type: string
        - in: query
          name: page_size
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: total
          schema:
            type: boolean
      responses:
        200:
          description: Users matching the search
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageOfUsers' }
...
  /api/exports/:
    get:
      tags: [ Export ]
      operationId: getAllExports
      description: >
        Get all exports regardless of user.
      summary: Get a list of all exports.
      responses:
        200:
          description: all exports.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Export' }
    post:
      tags: [ Export ]
      operationId: export
      description: >
        Export Mage data (GeoJSON, KML, GeoPackage, CSV) in the background. The requesting user must have
        `READ_OBSERVATION_ALL` permission or `READ_OBSERVATION_EVENT` or an ACL entry in the
        event with `read`.
      summary: Export Mage data using a background task.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
        required: true
      responses:
        201:
          description: export id
          headers:
            location:
              description: The URI of the export download.
              schema:
                type: string
          content:
            text/plain:
              schema:
                type: string
...
  /api/exports/{exportId}:
    get:
      tags: [ Export ]
      operationId: getExportById
      description: >
        Download an export by id.
      summary: Download a completed export.
      parameters:
      - name: exportId
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: The exported data as a compressed file.
          content:
            application/zip:
              schema:
                type: object
components:
  securitySchemes:
    authenticationToken:
      type: http
      description: >
        Token generated after successful authentication.  The client will exchange
        this token with an approved device for and API token.
      scheme: bearer
      bearerFormat: JWT
    mageToken:
      type: http
      description: >
        Mage generates an API token when a client authenticates
        initially.  The client can use the token for subsequent requests until
        it expires.
      scheme: bearer
  schemas:
...
    ColorHex:
      description: >
        This is a 3 or 4 byte hexadecimal string prefixed with '#' representing
        an RGB or RGBA color value, e.g. '#ff4545',
      type: string
      pattern: '#([a-fA-F0-9]{2}){3,4}'
...
    User:
      description: >
        The `User` schema defines the structure of persisted user documents as
        they exist in the database and as the server returns them in various
        responses.
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        status:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        lastUpdated:
          type: string
          format: date-time
        active:
          type: boolean
        enabled:
          type: boolean
        avatarUrl:
          type: string
        iconUrl:
          type: string
        authentication:
          $ref: '#/components/schemas/AuthStatus'
        role:
          $ref: '#/components/schemas/Role'
        roleId:
          $ref: '#/components/schemas/Role/properties/id'
        icon:
          $ref: '#/components/schemas/UserIcon'
        phones:
          type: array
          items:
            $ref: '#/components/schemas/Phone'
        recentEventIds:
          type: array
          items:
            $ref: '#/components/schemas/Event/properties/id'
      oneOf:
        - { type: object, required: [ role ], not: { type: object, required: [ roleId ] } }
        - { type: object, required: [ roleId ], not: { type: object, required: [ role ] } }
        - { type: object, not: { type: object, required: [ role, roleId ] } }
    UserRef:
      description: >
        `UserRef` is an object that references an existing user document with
        the required `id` property.  Mage ignores any properties other than
        `id`.
      type: object
      allOf:
        - { $ref: '#/components/schemas/User' }
        - { type: object, required: [ id ] }
    UserIn:
      type: object
      properties:
        username: { $ref: '#/components/schemas/User/properties/username' }
        displayName: { $ref: '#/components/schemas/User/properties/displayName' }
        email: { $ref: '#/components/schemas/User/properties/email' }
        phone:
          type: string
          format: phone
        roleId: { $ref: '#/components/schemas/User/properties/roleId' }
        iconMetadata:
          type: object
          properties:
            type: { $ref: '#/components/schemas/UserIcon/properties/type' }
            color: { $ref: '#/components/schemas/UserIcon/properties/color' }
            text: { $ref: '#/components/schemas/UserIcon/properties/text' }
        icon:
          description: >
            This key is valid only when encoded as `multipart/form-data`.
            This is a binary image file attachment.
          type: string
          format: binary
        avatar:
          description: >
            This key is valid only when encoded as `multipart/form-data`.
            This is a binary image file attachment.
          type: string
          format: binary
      allOf:
        - { $ref: '#/components/schemas/UserPasswordPair' }
...
    UserIcon:
      description: >
        A user icon is the small image that appears on a map to show the user's
        location.
      type: object
      properties:
        type:
          description: The origin of the icon
          type: string
          enum:
            - create
            - upload
            - none
        color:
          description: Color is only applicable for `create` type icons.
          $ref: '#/components/schemas/ColorHex'
        text:
          description: >
            Two-letter text label that appears on the map icon; only applies to
            `create` type icons
          type: string
        contentType:
          description: The MIME type of the icon image
          type: string
        size:
          description: The image size in bytes
          type: integer
    UserPasswordPair:
      type: object
      properties:
        password:
          type: string
        passwordconfirm:
          type: string
      oneOf:
        - { type: object, required: [ password, passwordconfirm ] }
        - { type: object, not: { type: object, required: [ password, passwordconfirm ] } }
    Authentication:
      type: object
      required:
        - user
        - token
      properties:
        user:
          $ref: '#/components/schemas/UserRef'
        token:
          type: string
    AuthStatus:
      type: object
      required:
        - type
      properties:
        type:
          type: string
        security:
          type: object
          properties:
            locked:
              type: boolean
            lockedUntil:
              type: string
              format: date-time
            invalidLoginAttempts:
              type: number
            numberOfTimesLocked:
              type: number
...
    DeviceIn:
      type: object
      required: [ uid ]
      properties:
        uid:
          type: string
        name:
          type: string
        description:
          type: string
        appVersion:
          type: string
        userId:
          $ref: '#/components/schemas/User/properties/id'
...
    Phone:
      type: object
      properties:
        type:
          description: >
            The type of phone number currently defaults to `Main` and is
            static.
          type: string
        number:
          type: string
...
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - description
        - permissions
...
    Export:
      required:
        - type
      type: object
      properties:
        _id:
          type: string
        userId:
          $ref: '#/components/schemas/User/properties/id'
        physicalPath:
          type: string
        filename:
          type: string
        exportType:
          type: string
          enum: [ geojson, kml, geopackage, csv ]
        location:
          type: string
        status:
          type: string
        options:
          type: string
...
    UserSearchResult:
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
        active:
          type: boolean
        enabled:
          type: boolean
        allPhones:
          type: string
    PageOfUsers:
      required:
        - totalCount
        - pageSize
        - pageIndex
        - items
      properties:
        totalCount:
          type: integer
        pageSize:
          type: integer
        pageIndex:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserSearchResult'
...
    ExportRequest:
      required:
        - exportType
        - eventId
        - observations
        - locations
      properties:
        exportType:
          $ref: '#/components/schemas/Export/properties/exportType'
        eventId:
          $ref: '#/components/schemas/Event/properties/id'
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        observations:
          type: boolean
        locations:
          type: boolean
        attachments:
          type: boolean
        favorites:
          type: boolean
        important:
          type: boolean
...